+++ example/android/app/src/main/kotlin/com/example/eps_payment_gateway_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.eps_payment_gateway_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()